Support search in apps and multi-cluster apps list page
