+++ app/apps-tab/index/template.hbs
@@ -8,6 +8,10 @@
 </section>
 
 <section class="pl-0 pr-0">
+  <div class="row mb-15">
+    {{search-text searchText=searchText}}
+  </div>
+
   {{#each filteredApps as |group|}}
     <div class="row flex">
       {{#each group as |ns|}}
@@ -18,7 +22,11 @@
     </div>
   {{else}}
     <div class="text-center text-muted">
-      {{t 'nav.apps.noData'}}
+      {{#if searchText}}
+        {{t "nav.apps.noMatch"}}
+      {{else}}
+        {{t "nav.apps.noData"}}
+      {{/if}}
     </div>
   {{/each}}
 </section>
