+++ app/components/namespace-table/component.js
@@ -2,7 +2,7 @@ import Component from '@ember/component';
 import layout from './template';
 import { inject as service } from '@ember/service';
 import { get, computed } from '@ember/object';
-import { matches } from 'shared/components/sortable-table/component';
+import { filter } from 'ui/utils/search-text';
 
 const headers = [
   {
@@ -45,40 +45,6 @@ export default Component.extend({
   ],
 
   projectsWithoutNamespace: computed('projectsWithoutNamespaces.[]', 'searchText', function() {
-    let searchText     = (get(this, 'searchText') || '').trim().toLowerCase();
-    let out            = get(this, 'projectsWithoutNamespaces').slice();
-    let searchFields   = ['displayName'];
-
-    if ( searchText.length ) {
-      let searchTokens = searchText.split(/\s*[, ]\s*/);
-
-      for ( let i = out.length - 1 ; i >= 0 ; i-- ) {
-        let hits      = 0;
-        let row       = out[i];
-        let mainFound = true;
-
-        for ( let j = 0 ; j < searchTokens.length ; j++ ) {
-          let expect = true;
-          let token  = searchTokens[j];
-
-          if ( token.substr(0, 1) === '!' ) {
-            expect = false;
-            token  = token.substr(1);
-          }
-
-          if ( token && matches(searchFields, token, row) !== expect ) {
-            mainFound = false;
-
-            break;
-          }
-        }
-
-        if ( !mainFound && hits === 0 ) {
-          out.removeAt(i);
-        }
-      }
-    }
-
-    return out
+    return filter(get(this, 'projectsWithoutNamespaces').slice(), get(this, 'searchText'), ['displayName']);
   }),
 });
