+++ lib/global-admin/addon/multi-cluster-apps/index/controller.js
@@ -3,6 +3,7 @@ import Controller from '@ember/controller';
 import C from 'ui/utils/constants';
 import { computed, get, observer } from '@ember/object';
 import { once } from '@ember/runloop';
+import { filter } from 'ui/utils/search-text';
 
 export default Controller.extend({
   prefs:   service(),
@@ -14,10 +15,11 @@ export default Controller.extend({
     once(() => this.get('catalog').fetchAppTemplates(get(this, 'model.apps')));
   }),
 
-  filteredApps: computed('model.apps.@each.{type,isFromCatalog,state}', function() {
+  filteredApps: computed('model.apps.@each.{type,isFromCatalog,state}', 'searchText', function() {
     let apps = get(this, 'model.apps').filter((ns) => !C.REMOVEDISH_STATES.includes(get(ns, 'state')));
 
     apps = apps.sortBy('displayName');
+    apps = filter(apps, get(this, 'searchText'));
 
     const group = [];
     let dataIndex = 0;
