+++ lib/shared/addon/components/search-text/template.hbs
@@ -1,0 +1,17 @@
+<div class="pull-right search-group input-group">
+  {{input
+    value=searchText
+    aria-title=(t "generic.search")
+    type="search" class="input-sm pull-right"
+    placeholder=(t "generic.search")
+  }}
+  {{#if searchText}}
+    <span class="input-group-btn">
+      <button
+        class="btn bg-transparent text-info pl-10 pr-10" {{action "clearSearch"}}
+      >
+        <i class="icon icon-close"/>
+      </button>
+    </span>
+  {{/if}}
+</div>
