+++ lib/shared/addon/components/sortable-table/component.js
@@ -11,6 +11,7 @@ import { isArray } from '@ember/array';
 import { observer } from '@ember/object'
 import { run } from '@ember/runloop';
 import { isAlternate, isMore, isRange } from 'shared/utils/platform';
+import { filter } from 'ui/utils/search-text';
 
 function toggleInput(node, on) {
   let id = get(node, 'id');
@@ -32,70 +33,6 @@ function toggleInput(node, on) {
       }
     }
   }
-}
-
-export function matches(fields, token, item) {
-  let tokenMayBeIp = /^[0-9a-f\.:]+$/i.test(token);
-
-  for ( let i = 0 ; i < fields.length ; i++ ) {
-    let field = fields[i];
-
-    if ( field ) {
-      // Modifiers:
-      //  id: The token must match id format (i.e. 1i123)
-      let idx = field.indexOf(':');
-      let modifier = null;
-
-      if ( idx > 0 ) {
-        modifier = field.substr(idx + 1);
-        field = field.substr(0, idx);
-      }
-
-      let val = get(item, field);
-
-      if ( val === undefined ) {
-        continue;
-      }
-
-      val = (`${ val }`).toLowerCase();
-      if ( !val ) {
-        continue;
-      }
-
-      switch ( modifier ) {
-      case 'exact':
-        if ( val === token ) {
-          return true;
-        }
-
-        break;
-      case 'ip':
-        if ( tokenMayBeIp ) {
-          let re = new RegExp(`(?:^|\.)${  token  }(?:\.|$)`);
-
-          if ( re.test(val) ) {
-            return true;
-          }
-        }
-
-        break;
-      case 'prefix':
-        if ( val.indexOf(token) === 0) {
-          return true;
-        }
-
-        break;
-      default:
-        if ( val.indexOf(token) >= 0) {
-          return true;
-        }
-
-        break;
-      }
-    }
-  }
-
-  return false;
 }
 
 export default Component.extend(Sortable, StickyHeader, {
@@ -446,78 +383,17 @@ export default Component.extend(Sortable, StickyHeader, {
   }),
 
   filtered: computed('arranged.[]', 'searchText', function() {
-    let out            = get(this, 'arranged').slice();
-    let searchFields   = get(this, 'searchFields');
-    let searchText     = (get(this, 'searchText') || '').trim().toLowerCase();
-    let subSearchField = get(this, 'subSearchField');
-    let subFields      = get(this, 'subFields');
-    let subMatches     = null;
-
-    if ( searchText.length ) {
-      subMatches = {};
-
-      let searchTokens = searchText.split(/\s*[, ]\s*/);
-
-      for ( let i = out.length - 1 ; i >= 0 ; i-- ) {
-        let hits      = 0;
-        let row       = out[i];
-        let mainFound = true;
-
-        for ( let j = 0 ; j < searchTokens.length ; j++ ) {
-          let expect = true;
-          let token  = searchTokens[j];
-
-          if ( token.substr(0, 1) === '!' ) {
-            expect = false;
-            token  = token.substr(1);
-          }
-
-          if ( token && matches(searchFields, token, row) !== expect ) {
-            mainFound = false;
-
-            break;
-          }
-        }
-
-        if ( subFields && subSearchField) {
-          let subRows = (row.get(subSearchField) || []);
-
-          for ( let k = subRows.length - 1 ; k >= 0 ; k-- ) {
-            let subFound = true;
-
-            for ( let l = 0 ; l < searchTokens.length ; l++ ) {
-              let expect = true;
-              let token  = searchTokens[l];
-
-              if ( token.substr(0, 1) === '!' ) {
-                expect = false;
-                token  = token.substr(1);
-              }
-
-              if ( matches(subFields, token, subRows[k]) !== expect ) {
-                subFound = false;
-
-                break;
-              }
-            }
-
-            if ( subFound ) {
-              hits++;
-            }
-          }
-
-          subMatches[row.get('id')] = hits;
-        }
-
-        if ( !mainFound && hits === 0 ) {
-          out.removeAt(i);
-        }
-      }
-    }
+    const { matches, subMatches } = filter(
+      get(this, 'arranged').slice(),
+      get(this, 'searchText'),
+      get(this, 'searchFields'),
+      get(this, 'subFields'),
+      get(this, 'subSearchField')
+    );
 
     set(this, 'subMatches', subMatches);
 
-    return out;
+    return matches;
   }),
 
   indexFrom: computed('page', 'perPage', function() {
