+++ lib/shared/app/components/search-text/component.js
@@ -1,0 +1,1 @@
+export { default } from 'shared/components/search-text/component';
