+++ lib/shared/app/utils/search-text.js
@@ -1,0 +1,1 @@
+export { filter } from 'shared/utils/search-text';
