+++ translations/en-us.yaml
@@ -1573,6 +1573,7 @@ globalDnsPage:
 multiClusterAppsPage:
   header: Multi-Cluster Apps
   noData: There are no multi-cluster apps launched
+  noMatch: No multi-cluster apps match the current search
   error:
     appData: Error loading global app data
 
@@ -7764,6 +7765,7 @@ nav:
     tab: Apps
     apps: Apps
     noData: There are no apps launched.
+    noMatch: No apps match the current search
     launch: Launch
     manage: Manage Catalogs
   infra:
